+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/arg_parser.go
@@ -246,7 +246,7 @@ func ParseNonEmptyString(argName string, argValue starlark.Value) (string, *star
 	return strArgValue, nil
 }
 
-func ParseArtifactId(artifactUuidArgName string, artifactIdStr starlark.String) (enclave_data_directory.FilesArtifactUUID, *startosis_errors.InterpretationError) {
+func ParseArtifactId(artifactUuidArgName string, artifactIdStr starlark.String) (enclave_data_directory.FilesArtifactID, *startosis_errors.InterpretationError) {
 	artifactId, interpretationErr := safeCastToString(artifactIdStr, artifactUuidArgName)
 	if interpretationErr != nil {
 		return "", interpretationErr
@@ -254,7 +254,7 @@ func ParseArtifactId(artifactUuidArgName string, artifactIdStr starlark.String) 
 	if len(artifactId) == 0 {
 		return "", startosis_errors.NewInterpretationError("Artifact Uuid can't be empty for argument '%s'", artifactUuidArgName)
 	}
-	return enclave_data_directory.FilesArtifactUUID(artifactId), interpretationErr
+	return enclave_data_directory.FilesArtifactID(artifactId), interpretationErr
 }
 
 func ParseTemplatesAndData(templatesAndData *starlark.Dict) (map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData, *startosis_errors.InterpretationError) {